+++ src/Types/UserRoles.php
@@ -4,8 +4,9 @@ namespace App\Types;
 
 enum UserRoles: string
 {
-    case ROLE_USER = 'USER';
-    case ROLE_ADMIN = 'ADMIN';
+    case ROLE_USER = 'ROLE_USER';
+    case ROLE_ADMIN = 'ROLE_ADMIN';
+    case ROLE_API_USER = 'ROLE_API_USER';
 
     public static function array(): array
     {